+++ src/types/index.ts
@@ -1,3 +1,3 @@
 export * from './BMFont';
 export * from './BMFontUtil';
-export * from './Layouts';
+export * from './LayoutTypes';
